mv asset spec tests to test_asset_decorator, use @dlt_asset encapsulation
